Separated asset json data loading logic into a separate class.
